// src/app/modules/PIS/models/IRDA/list-pasien-rawat-darurat-anonim.model.ts
export interface IDaftarPasienRawatDaruratAnonimModel {
    id_register: number
    no_register: string
    id_person: number
    no_rekam_medis: string
    jenis_rawat: string
    jam_masuk: string
    nama_pasien: string
    nama_dokter: string
    gender: string
    keterangan_diagnosa: string
    keluhan: string
}